git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@37897 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/octagon/values/OctagonDoubleValue.java
@@ -50,7 +50,7 @@ public class OctagonDoubleValue extends OctagonNumericValue<Double> {
 
   @Override
   public OctagonNumericValue add(OctagonNumericValue pVal) {
-    return pVal.add(value.doubleValue());
+    return pVal.add(value);
   }
 
   @Override
